+++ packages/form/src/module.ts
@@ -92,6 +92,7 @@ import { TimeWidget } from './widgets/time/time.widget';
 import { TransferWidget } from './widgets/transfer/transfer.widget';
 import { TreeSelectWidget } from './widgets/tree-select/tree-select.widget';
 import { UploadWidget } from './widgets/upload/upload.widget';
+import { NzWidgetRegistry } from './widgets';
 
 const WIDGETS = [
   ObjectWidget,
@@ -136,7 +137,7 @@ export class YelonFormModule {
           useClass: AjvSchemaValidatorFactory,
           deps: [YunzaiConfigService, NgZone]
         },
-        { provide: WidgetRegistry, useClass: WidgetRegistry }
+        { provide: WidgetRegistry, useClass: NzWidgetRegistry }
       ]
     };
   }