+++ scripts/ci/helper.sh
@@ -1,4 +1,3 @@
-#!/usr/bin/env bash
 rm -rf archive-docs
 git clone https://github.com/hbyunzai/archive-docs.git
 cp -r archive-docs/issue-helper src/dist/browser/issue-helper